+++ src/inode.rs
@@ -76,99 +76,99 @@ pub fn inode_find(path: [char; directory::DIR_NAME], mmap: &memmap2::MmapMut) ->
 	return alloc_inode(path, data);
 }*/
 
-pub fn inode_deserialize(mmap: &[u8], num: i32) -> Inode {
+pub fn inode_deserialize(mmap: &[i8], num: i32) -> Inode {
 	let offset: usize = (num as usize) * std::mem::size_of::<Inode>();
 	
 	let mut data: [u8; 4] = [0; 4];
 	let mut data16: [u8; 2] = [0; 2];
 	for i in 3..=0 {	// Endian: big/little
-		data[i] = mmap[INS+offset+i..INS+offset+i+1][0];
+		data[i] = mmap[INS+offset+i..INS+offset+i+1][0] as u8;
 	}
 	
 	let refs: u32 = u32::from_ne_bytes(data);
 	
 	for i in 3..=0 {
-		data[i] = mmap[INS+offset+4+i..INS+offset+4+i+1][0];
+		data[i] = mmap[INS+offset+4+i..INS+offset+4+i+1][0] as u8;
 	}
 	let mode: u32 = u32::from_ne_bytes(data);
 	
 	for i in 3..=0 {
-		data16[i] = mmap[INS+offset+8+i..INS+offset+8+i+1][0];
+		data16[i] = mmap[INS+offset+8+i..INS+offset+8+i+1][0] as u8;
 	}
 	let size0: u16 = u16::from_ne_bytes(data16);
 	
 	for i in 1..=0 {
-		data16[i] = mmap[INS+offset+10+i..INS+offset+10+i+1][0];
+		data16[i] = mmap[INS+offset+10+i..INS+offset+10+i+1][0] as u8;
 	}
 	let size1: u16 = u16::from_ne_bytes(data16);
 	
 	let sizes = [size0, size1];
 	
 	for i in 1..=0 {
-		data16[i] = mmap[INS+offset+12+i..INS+offset+12+i+1][0];
+		data16[i] = mmap[INS+offset+12+i..INS+offset+12+i+1][0] as u8;
 	}
 	let ptrs0: u16 = u16::from_ne_bytes(data16);
 	
 	for i in 1..=0 {
-		data16[i] = mmap[INS+offset+14+i..INS+offset+14+i+1][0];
+		data16[i] = mmap[INS+offset+14+i..INS+offset+14+i+1][0] as u8;
 	}
 	let ptrs1: u16 = u16::from_ne_bytes(data16);
 	
 	let ptrs = [ptrs0, ptrs1];
 	
 	for i in 3..=0 {
-		data[i] = mmap[INS+offset+16+i..INS+offset+16+i+1][0];
+		data[i] = mmap[INS+offset+16+i..INS+offset+16+i+1][0] as u8;
 	}
 	let iptr: i32 = i32::from_ne_bytes(data);
 	
 	for i in 3..=0 {
-		data[i] = mmap[INS+offset+20+i..INS+offset+20+i+1][0];
+		data[i] = mmap[INS+offset+20+i..INS+offset+20+i+1][0] as u8;
 	}
 	let inum: i32 = i32::from_ne_bytes(data);
 	
 	return Inode { refs: refs, mode: mode, size: sizes, ptrs: ptrs, iptr: iptr, inum: inum };
 }
 
-pub fn inode_serialize(mmap: &mut [u8], d: Inode) -> i32 {
+pub fn inode_serialize(mmap: &mut [i8], d: Inode) -> i32 {
 	let offset: usize = (d.inum as usize) * std::mem::size_of::<Inode>();
 	
 	for i in 3..=0 {
-		mmap[INS+offset+i..INS+offset+i+1][0] = d.refs.to_be_bytes()[i];
+		mmap[INS+offset+i..INS+offset+i+1][0] = d.refs.to_be_bytes()[i] as i8;
 	}
 	
 	for i in 3..=0 {
-		mmap[INS+offset+4+i..INS+offset+4+i+1][0] = d.mode.to_be_bytes()[i];
+		mmap[INS+offset+4+i..INS+offset+4+i+1][0] = d.mode.to_be_bytes()[i] as i8;
 	}
 	
 	for i in 2..=0 {
-		mmap[INS+offset+8+i..INS+offset+8+i+1][0] = d.size[0].to_be_bytes()[i];
+		mmap[INS+offset+8+i..INS+offset+8+i+1][0] = d.size[0].to_be_bytes()[i] as i8;
 	}
 	
 	for i in 2..=0 {
-		mmap[INS+offset+10+i..INS+offset+10+i+1][0] = d.size[1].to_be_bytes()[i];
+		mmap[INS+offset+10+i..INS+offset+10+i+1][0] = d.size[1].to_be_bytes()[i] as i8;
 	}
 	
 	for i in 2..=0 {
-		mmap[INS+offset+12+i..INS+offset+12+i+1][0] = d.ptrs[0].to_be_bytes()[i];
+		mmap[INS+offset+12+i..INS+offset+12+i+1][0] = d.ptrs[0].to_be_bytes()[i] as i8;
 	}
 	
 	for i in 2..=0 {
-		mmap[INS+offset+14+i..INS+offset+14+i+1][0] = d.ptrs[1].to_be_bytes()[i];
+		mmap[INS+offset+14+i..INS+offset+14+i+1][0] = d.ptrs[1].to_be_bytes()[i] as i8;
 	}
 	
 	for i in 3..=0 {
-		mmap[INS+offset+16+i..INS+offset+16+i+1][0] = d.iptr.to_be_bytes()[i];
+		mmap[INS+offset+16+i..INS+offset+16+i+1][0] = d.iptr.to_be_bytes()[i] as i8;
 	}
 	
 	for i in 3..=0 {
-		mmap[INS+offset+20+i..INS+offset+20+i+1][0] = d.inum.to_be_bytes()[i];
+		mmap[INS+offset+20+i..INS+offset+20+i+1][0] = d.inum.to_be_bytes()[i] as i8;
 	}
 	
 	return d.inum;
 }
 
-pub fn inode_read(d: Inode, mmap: &[u8]) -> (Vec<u8>, i32) {
-	let mut c: Vec<u8> = vec![];
+pub fn inode_read(d: Inode, mmap: &[i8]) -> (Vec<i8>, i32) {
+	let mut c: Vec<i8> = vec![];
 	for i in 0..=d.size[0]-1 {
 		c.push(mmap[INS+(d.ptrs[0] as usize)..INS+(d.ptrs[0] as usize)+1][0]);
 	}
